Move controllers to their own package
